Issue 60: Clean up new clang-tidy issues and fix doxygen output (#62)

+++ reference/src/taimpl/src/ta_sa_crypto_sign.c
@@ -1,4 +1,4 @@
-/**
+/*
  * Copyright 2020-2023 Comcast Cable Communications Management, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -88,12 +88,12 @@ static sa_status ta_sa_crypto_sign_ecdsa(
         }
 
         if (out == NULL) {
-            *out_length = header->size * 2;
+            *out_length = (size_t) header->size * 2;
             status = SA_STATUS_OK;
             break;
         }
 
-        if (*out_length < header->size * 2) {
+        if (*out_length < (size_t) header->size * 2) {
             ERROR("Invalid out_length");
             status = SA_STATUS_INVALID_PARAMETER;
             break;
